+++ src/main/java/it/gov/pagopa/pu/workflow/ingestionflow/PaymentsReportingIngestionWF.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.pu.workflow;
+package it.gov.pagopa.pu.workflow.ingestionflow;
 
 import io.temporal.workflow.WorkflowInterface;
 import io.temporal.workflow.WorkflowMethod;
@@ -12,7 +12,7 @@ public interface PaymentsReportingIngestionWF {
 
   /**
    * Workflow method for the Payments Reporting Ingestion Workflow
-   * @param fileId
+   * @param ingestionFlowFileId
    * */
   @WorkflowMethod
   public void ingest(Long ingestionFlowFileId);
